+++ src/tasks/golang.rs
@@ -1,8 +1,8 @@
-use lib::{
+use crate::lib::{
     env::Exports,
     task::{self, Status, Task},
 };
-use utils::{
+use crate::utils::{
     self,
     fs::mktemp,
     golang::{arch, bin_dir, os},