+++ src/tasks/rustc.rs
@@ -1,6 +1,6 @@
 use regex;
 
-use lib::{
+use crate::lib::{
     rust,
     task::{self, Status, Task},
 };